# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel API",
        "description": "API para gerenciamento de solicitações de viagem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria uma nova conta de usuário e retorna o token de acesso",
                "operationId": "460b8091c3f2976ca951cacd50b7c203",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome completo do usuário",
                                        "type": "string",
                                        "example": "Luis Henrique"
                                    },
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "format": "email",
                                        "example": "luis@example.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário (mínimo 8 caracteres, deve conter maiúscula, minúscula, número e caractere especial)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "MinhaSenh@123"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmação da senha",
                                        "type": "string",
                                        "format": "password",
                                        "example": "MinhaSenh@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário registrado e logado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Luis Henrique"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "luis@example.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "user"
                                                        },
                                                        "role_label": {
                                                            "type": "string",
                                                            "example": "Usuário"
                                                        },
                                                        "email_verified_at": {
                                                            "type": "string",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2024-01-01 10:00:00"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "example": "2024-01-01 10:00:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dados de validação inválidos"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "O email já está sendo usado por outro usuário."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno do servidor"
                                        },
                                        "error_code": {
                                            "type": "string",
                                            "example": "REGISTRATION_ERROR"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Fazer logout",
                "description": "Invalida o token JWT do usuário autenticado",
                "operationId": "8f22b4b1d4d13235593868f409cf2395",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token inválido ou expirado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro no logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro no logout"
                                        },
                                        "error_code": {
                                            "type": "string",
                                            "example": "LOGOUT_ERROR"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}